Fix web build size_t/uint64_t

+++ src/engine/scene_binary_format.h
@@ -6,5 +6,5 @@ constexpr uint8_t MAX_SCENE_NAME_SIZE = 64;
 
 struct sSceneBinaryHeader {
     uint8_t version = 1;
-    size_t node_count = 0;
+    uint64_t node_count = 0;
 };
