+++ src/framework/nodes/node_binary_format.h
@@ -5,5 +5,5 @@
 constexpr uint8_t MAX_NODE_NAME_SIZE = 64;
 
 struct sNodeBinaryHeader {
-    size_t children_count = 0;
+    uint64_t children_count = 0;
 };
